Use healthy instances for asg replacement
Unhealthy instances != 0 is not synonymous with healthy instances > 0.
For traffic to be routed through LBs (v1 or v2), there needs to be at
least one healhty instance.  This patch uses the condition that the
number of healthy instances in Load Balancers must equal the new ASG
desired capacity, and the desired capacity must not equal 0. Two
conditions that must be true for LBs to route traffic.